Update fabric.mod.json
In an attempt to resolve an issue caused when moving from 1.18.1 to 1.18.2, I first tried to make the logging check completely optional with mixins. When that didn't work, I said fuck it and removed the check and related events entirely. However, I forgot to actually remove the reference to the now-nonexistant mixin file.

This corrects that issue.